requirements.txt
Django==5.0.3
easy-thumbnails==2.8.5
Pillow==10.0.0
channels==4.1.0
daphne==4.1.2
channels-redis==4.2.0
factory-boy==3.3.1
psycopg2==2.9.10